+++ R/ggplotStuff.R
@@ -1,12 +1,12 @@
 
-.gumBreaks_p <- c(0.5, 0.2, 0.1, 0.05, 0.02, 0.01,  0.005, 0.002, 0.001)
+.gumBreaks_p <- c(0.1, 0.01, 0.001, 0.0001)
 
 ##' @importFrom scales trans_new
 .gumbel_trans_p <- trans_new(name = "gumbel",
                              transform = function(p) -log(-log(1 - p)),
                              inverse = function(y) 1 - exp(-exp(-y)),
                              breaks = myBreaks <- c(0.5, 0.2, 0.1,
-                                 0.05, 0.02, 0.01,  0.005, 0.002, 0.001),
+                                 0.05, 0.02, 0.01,  0.005, 0.002, 0.001, 0.0005, 0.0001),
                              domain = c(0, 1))
 
 .gumBreaks_m <- c(2, 5, 10, 20, 50, 100, 200, 500, 1000)
@@ -409,17 +409,60 @@ autoplot.simulate.pgpTList <- function(object, ...) {
 ##' @export
 ##' @method autoplot quantile.pgpTList
 ##' 
-autoplot.quantile.pgpTList <- function(object, ...) {
+autoplot.quantile.pgpTList <- function(object, facet = TRUE,
+                                       ...) {
     
-    g <- ggplot(data = object) +
-        geom_line(mapping = aes_string(x = "ProbExc", y = "Quant",
-                                       colour = "tau"))
+    g <- ggplot(data = object)
+    
+        ## geom_line(mapping = aes_string(x = "ProbExc", y = "Quant",
+        ##                               colour = "tau"))
     
     g <- g + scale_x_continuous(trans = .gumbel_trans_p,
                                 breaks = .gumBreaks_p,
                                 minor_breaks = .gumBreaks_p) +
         xlab("Prob. of exceedance") + ylab("Quantile") +
-        ggtitle("Tail distribution for the maximum")
+        ggtitle(sprintf(paste("Tail distribution for the maximum with %2.0f %%",
+                        "confidence interval"),
+                        100 * attr(object, "level")))
+       
+    if (facet) {
+        g <- g +
+            geom_ribbon(
+                data = object,
+                mapping = aes_string(x = "ProbExc", ymin = "L", ymax = "U"),
+                fill = "SteelBlue3", alpha = 0.3)
+        g <- g +
+            geom_line(
+                data = object,
+                mapping = aes_string(x = "ProbExc", y = "L"),
+                colour = "SteelBlue3")
+        g <- g +
+            geom_line(
+                data = object,
+                mapping = aes_string(x = "ProbExc", y = "U"),
+                colour = "SteelBlue4")
+        g <- g + geom_line(mapping = aes_string(x = "ProbExc", y = "Quant",
+                                                colour = "tau"), size = 0.6)
+        
+        g <- g + scale_colour_brewer(palette = "Set2")
+        g <- g + facet_wrap(tau ~ ., labeller = label_both)
+    } else {
+        g <- g +
+            geom_line(
+                data = object,
+                mapping = aes_string(x = "ProbExc", y = "L", colour = "tau"),
+                linetype = "dashed")
+        g <- g +
+            geom_line(
+                data = object,
+                mapping = aes_string(x = "ProbExc", y = "U", colour = "tau"),
+                linetype = "dashed")
+        g <- g + geom_line(mapping = aes_string(x = "ProbExc", y = "Quant",
+                                                colour = "tau"), size = 0.6)
+        
+        g <- g + scale_colour_brewer(palette = "Set2")
+    }
+   
     g
 
 }